Reimplement KeyCombo#updateState
The old implementation was originally designed to optimize for tick
batching so that state per key combo was only calculated once per tick.
while this was good for performance, it made it very difficult for
users to interact with events (for example calling preventDefault).

The new implementation is a second take after learning from the limits
and issues encountered from the first. The largest problem solved is
sequence advancement happens once all keys are released. Random keys
can no longer be pressed in-between sequences (this was never suppose)
to work this way.